ci: freshen
CI is stale enough to not be able to run.

Circle CI's clojure docker images are not as well maintained as they used to be
so switch to clojure docker image.

Attempt the correct incantation to get Planck installed.

Other minor tweaks, like adding step names and dumping tools versions